Use .env_test for non-master/main branches
Updated environment variable loading in globals.js to use .env_test when the current Git branch is not master or main. Added helper functions in helpersFunctions.js to detect the current Git branch and determine if it is master/main. This allows for safer environment separation between production and test branches.